solutionfabric/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-celery-beat==2.3.0
django-timezone-field==5.0
djangorestframework==3.13.1
idna==3.3
kombu==5.2.4
prompt-toolkit==3.0.30
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
